---

The attached commit changes the fSecs field of plUnifiedTime from UInt32 to time_t. This fixes various date formatting problems, e.g. missing log timestamps or output of the Net.GetServerTime console command, when building on a system where time_t is 64-bit (e.g. Visual Studio 2010, and possibly also 2008, though I didn’t test that), and, as a bonus, extends the range past 2038 on such systems.

The wire protocol is left at 32 bit for now, we might change that to 64 when other reasons to break compatibility have accumulated.

Avoiding unnecessary back-and-forth type conversions makes the change ripple out into several other files. Changes have been tested as far as I easily could, and introduce no new compiler warnings.

The minimal-invasive way of fixing only the date formatting, should you prefer that, is in branch [cwalther:timet-minimal](https://github.com/cwalther/Plasma/branches/timet-minimal) (9b54fb05c9471c1b3317a5f8ed6cd4c11977e9e1).

+++ Sources/Plasma/FeatureLib/pfPython/cyMiscGlue.cpp
@@ -59,12 +59,12 @@ PYTHON_GLOBAL_METHOD_DEFINITION_NOARGS(PtGetPrevAgeInfo, "Returns ptAgeInfoStruc
 
 PYTHON_GLOBAL_METHOD_DEFINITION_NOARGS(PtGetDniTime, "Returns current D'Ni time")
 {
-    return PyLong_FromUnsignedLong(cyMisc::GetDniTime());
+    return PyLong_FromUnsignedLong((unsigned long)cyMisc::GetDniTime());
 }
 
 PYTHON_GLOBAL_METHOD_DEFINITION_NOARGS(PtGetServerTime, "Returns the current time on the server (which is GMT)")
 {
-    return PyLong_FromUnsignedLong(cyMisc::GetServerTime());
+    return PyLong_FromUnsignedLong((unsigned long)cyMisc::GetServerTime());
 }
 
 PYTHON_GLOBAL_METHOD_DEFINITION(PtGMTtoDniTime, args, "Params: gtime\nConverts GMT time (passed in) to D'Ni time")
@@ -75,7 +75,7 @@ PYTHON_GLOBAL_METHOD_DEFINITION(PtGMTtoDniTime, args, "Params: gtime\nConverts G
         PyErr_SetString(PyExc_TypeError, "PtGMTtoDniTime expects a long");
         PYTHON_RETURN_ERROR;
     }
-    return PyLong_FromUnsignedLong(cyMisc::ConvertGMTtoDni(gtime));
+    return PyLong_FromUnsignedLong((unsigned long)cyMisc::ConvertGMTtoDni(gtime));
 }
 
 PYTHON_GLOBAL_METHOD_DEFINITION(PtGetClientName, args, "Params: avatarKey=None\nThis will return the name of the client that is owned by the avatar\n"
